+++ core/2d/line_dist.cpp
@@ -6,7 +6,7 @@ LineDist::LineDist(const dLine & _line): line(_line){
 
   double l=0;
   ls.push_back(0);
-  for (int j=1; j<_line.size(); j++){
+  for (size_t j=1; j<_line.size(); j++){
     dPoint p1 (_line[j-1]);
     dPoint p2 (_line[j]);
     if (p1 != p2) l+=pdist(p1,p2);
